refactor: remove Products component
