+++ src/components/Partner/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Image, { StaticImageData } from 'next/image'
-import useMatomoEcommerce from '../../../hooks/useMatomoEcommerce'
+import useMatomoEcommerce from '../../hooks/useMatomoEcommerce'
 
 export interface PartnerData {
   href: string;
